Rename packages for clarity
Also fixed incoherent DB schema
