+++ edtLHD/src/main/java/fr/univtln/lhd/exceptions/IdException.java
@@ -1,4 +1,4 @@
-package fr.univtln.lhd.exception;
+package fr.univtln.lhd.exceptions;
 
 public class IdException extends Exception{
     public IdException(String message) {
