+++ edtLHD/src/main/java/fr/univtln/lhd/model/entities/slots/Group.java
@@ -1,6 +1,6 @@
 package fr.univtln.lhd.model.entities.slots;
 
-import fr.univtln.lhd.model.entities.user.Student;
+import fr.univtln.lhd.model.entities.users.Student;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
