+++ edtLHD/src/test/java/fr/univtln/lhd/entities/slots/GroupTest.java
@@ -3,7 +3,7 @@ package fr.univtln.lhd.entities.slots;
 import static org.junit.jupiter.api.Assertions.*;
 
 import fr.univtln.lhd.model.entities.slots.Group;
-import fr.univtln.lhd.model.entities.user.Student;
+import fr.univtln.lhd.model.entities.users.Student;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
